Config: Abstract the config manager and parsed value
